Changed Mac OS X version to use CMake. Had to split into two projects. See pm_mac/README_MAC.txt

+++ portmidi/trunk/pm_java/jportmidi/JPortMidiApi.java
@@ -2,7 +2,7 @@ package jportmidi;
 
 public class JPortMidiApi {
     public static class PortMidiStream {
-        private int address;
+        private long address;
     }
     public static class PmEvent {
         public int message;
